+++ src/renderer/components/Report/YearlyCatByMonth.vue
@@ -1,24 +1,32 @@
 <template>
-<div>
-  <v-list-tile :class="rowColor(month)">
-    <v-list-tile-avatar>
-      <v-icon>{{ month.icon }}</v-icon>
-    </v-list-tile-avatar>
-    <v-layout row align-center>
-      <v-flex xs2><span class="subheading">{{ month.text }}</span></v-flex>
-      <v-flex xs10>
-        <v-layout row align-center>
-          <v-flex xs2 text-xs-center>{{ format.formatMoney(data.amount) }} / {{ format.formatMoney(Math.abs(data.budgetedAmount+0.0)) }}</v-flex>
-          <v-flex xs6><v-progress-linear v-model="progressPercent" height="20" :color="progressColor"></v-progress-linear></v-flex>
-          <v-flex xs1 text-xs-center>{{ progressPercent }}%</v-flex>
-        </v-layout>
-      </v-flex>
-    </v-layout>
-  </v-list-tile>
-</div>
+  <div>
+    <v-list-tile :class="rowColor(month)">
+      <v-list-tile-avatar>
+        <v-icon>{{ month.icon }}</v-icon>
+      </v-list-tile-avatar>
+      <v-layout row align-center>
+        <v-flex xs2>
+          <span class="subheading">{{ month.text }}</span>
+        </v-flex>
+        <v-flex xs10>
+          <v-layout row align-center>
+            <v-flex
+              xs2
+              text-xs-center
+            >{{ format.formatMoney(data.amount) }} / {{ format.formatMoney(Math.abs(data.budgetedAmount+0.0)) }}</v-flex>
+            <v-flex xs6>
+              <v-progress-linear v-model="progressPercent" height="20" :color="progressColor"></v-progress-linear>
+            </v-flex>
+            <v-flex xs1 text-xs-center>{{ progressPercent }}%</v-flex>
+          </v-layout>
+        </v-flex>
+      </v-layout>
+    </v-list-tile>
+  </div>
 </template>
 
 <script>
+import Constants from '../../lib/Constants'
 import Format from '../../lib/Format'
 
 export default {
@@ -44,13 +52,13 @@ export default {
       if (p === 0) {
         color = 'white'
       } else if (p <= 75) {
-        color = 'green accent-1'
+        color = Constants.COLORS.PROGRESS_GOOD
       } else if (p > 75 && p < 100) {
-        color = 'yellow accent-1'
+        color = Constants.COLORS.PROGRESS_WARN
       } else if (p === 100) {
-        color = 'green accent-3'
+        color = Constants.COLORS.PROGRESS_BULLSEYE
       } else {
-        color = 'red accent-2'
+        color = Constants.COLORS.PROGRESS_DANGER
       }
 
       return color
@@ -59,7 +67,7 @@ export default {
 
   methods: {
     rowColor: function (month) {
-      var color = month.value % 2 === 0 ? 'grey lighten-4' : 'grey lighten-2'
+      var color = month.value % 2 === 0 ? Constants.COLORS.GREY : Constants.COLORS.GREY_ALT
       return (color)
     }
   },