+++ src/renderer/router/index.js
@@ -21,14 +21,24 @@ export default new Router({
       component: require('@/components/Expenses').default
     },
     {
+      path: '/expenses/category/:category',
+      name: 'expenses-screen-category',
+      component: require('@/components/Expenses').default
+    },
+    {
       path: '/report/list',
       name: 'report-list',
       component: require('@/components/Report/Main').default
     },
     {
-      path: '/report/yearly',
-      name: 'report-yearly',
-      component: require('@/components/Report/Yearly').default
+      path: '/report/yearly-budget/:year?',
+      name: 'report-yearly-budget',
+      component: require('@/components/Report/YearlyBudget').default
+    },
+    {
+      path: '/report/multi-year-comparison',
+      name: 'report-multi-year-comparison',
+      component: require('@/components/Report/MultiYearComparison').default
     },
     {
       path: '*',